Split local dev flows
