+++ types/Yargs.ts
@@ -29,3 +29,5 @@ export type OverwriteArgs = Options & {
 export type StringArgType = Options & {
   type: 'string';
 };
+
+export type ProjectDevArgs = CommonArgs & ConfigArgs & EnvironmentArgs;
